Further improvements and cleanup
